Refactor extract module wordlist
